+++ client/src/components/Home/Home.tsx
@@ -3,8 +3,8 @@ import { Tag, Button, Table, Space } from "antd";
 import { SearchAutoComplete } from "./SearchAutoComplete";
 import { PlusOutlined, InfoCircleOutlined } from "@ant-design/icons";
 import { Link } from "react-router-dom";
-import { GetUser, GetUserVariables } from "../__generated__/getUser";
-import { GET_USER } from "../queries/GetUser.gql";
+import { GetUser, GetUserVariables } from "../../__generated__/getUser";
+import { GET_USER } from "../../__queries__/GetUser.gql";
 import { useQuery } from "@apollo/client";
 
 const columns = [
