+++ client/src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./i18n/i18next";
+import "./services/i18next";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import {
