+++ client/src/layout/AppSidebar.tsx
@@ -1,16 +1,15 @@
 import React from "react";
-import "../App.css";
 import { Layout, Menu } from "antd";
 import { TeamOutlined, BarChartOutlined } from "@ant-design/icons";
 import { useTranslation } from "react-i18next";
-import i18next from "../i18n/i18next";
+import i18next from "../services/i18next";
 import { Link } from "react-router-dom";
 import {
   GetUser,
   GetUserVariables,
   GetUser_user_teams_programs_datasets,
 } from "../__generated__/getUser";
-import { GET_USER } from "../queries/GetUser.gql";
+import { GET_USER } from "../__queries__/GetUser.gql";
 import { useQuery } from "@apollo/client";
 
 const { SubMenu } = Menu;
